refactor: 코스 application 패키지 분리
