+++ src/main/java/com/michelet/restaurantservice/course/application/command/CreateCourseCommand.java
@@ -1,8 +1,8 @@
-package com.michelet.restaurant.application.command;
+package com.michelet.restaurantservice.course.application.command;
 
-import com.michelet.restaurant.domain.model.CourseSessionType;
-import com.michelet.restaurant.domain.model.CourseStatus;
-import com.michelet.restaurant.presentation.dto.CreateCourseRequest;
+import com.michelet.restaurantservice.course.domain.model.CourseSessionType;
+import com.michelet.restaurantservice.course.domain.model.CourseStatus;
+import com.michelet.restaurantservice.presentation.dto.CreateCourseRequest;
 
 import java.util.List;
 import java.util.UUID;
