+++ src/main/java/com/michelet/restaurantservice/course/application/result/CourseListItemResult.java
@@ -1,8 +1,8 @@
-package com.michelet.restaurant.application.result;
+package com.michelet.restaurantservice.course.application.result;
 
-import com.michelet.restaurant.domain.model.CourseSessionType;
-import com.michelet.restaurant.domain.model.CourseStatus;
-import com.michelet.restaurant.domain.model.RestaurantCourse;
+import com.michelet.restaurantservice.course.domain.model.CourseSessionType;
+import com.michelet.restaurantservice.course.domain.model.CourseStatus;
+import com.michelet.restaurantservice.course.domain.model.RestaurantCourse;
 
 import java.util.List;
 import java.util.UUID;
