+++ src/main/java/com/michelet/restaurantservice/course/application/result/CourseResult.java
@@ -1,6 +1,6 @@
-package com.michelet.restaurant.application.result;
+package com.michelet.restaurantservice.course.application.result;
 
-import com.michelet.restaurant.domain.model.RestaurantCourse;
+import com.michelet.restaurantservice.course.domain.model.RestaurantCourse;
 
 import java.util.UUID;
