+++ src/main/java/com/michelet/restaurantservice/course/application/service/RestaurantCourseCommandService.java
@@ -1,17 +1,17 @@
-package com.michelet.restaurant.application.service.command;
+package com.michelet.restaurantservice.course.application.service;
 
-import com.michelet.restaurant.application.command.CreateCourseCommand;
-import com.michelet.restaurant.application.command.CreateCourseMenuCommand;
-import com.michelet.restaurant.application.result.CourseResult;
-import com.michelet.restaurant.domain.exception.CourseErrorCode;
-import com.michelet.restaurant.domain.exception.CourseException;
-import com.michelet.restaurant.domain.exception.RestaurantErrorCode;
-import com.michelet.restaurant.domain.exception.RestaurantException;
-import com.michelet.restaurant.domain.model.RestaurantCourse;
-import com.michelet.restaurant.domain.model.RestaurantCourseMenu;
-import com.michelet.restaurant.domain.repository.RestaurantCourseMenuRepository;
-import com.michelet.restaurant.domain.repository.RestaurantCourseRepository;
-import com.michelet.restaurant.domain.repository.RestaurantRepository;
+import com.michelet.restaurantservice.course.application.command.CreateCourseCommand;
+import com.michelet.restaurantservice.course.application.command.CreateCourseMenuCommand;
+import com.michelet.restaurantservice.course.application.result.CourseResult;
+import com.michelet.restaurantservice.domain.exception.CourseErrorCode;
+import com.michelet.restaurantservice.domain.exception.CourseException;
+import com.michelet.restaurantservice.domain.exception.RestaurantErrorCode;
+import com.michelet.restaurantservice.domain.exception.RestaurantException;
+import com.michelet.restaurantservice.course.domain.model.RestaurantCourse;
+import com.michelet.restaurantservice.course.domain.model.RestaurantCourseMenu;
+import com.michelet.restaurantservice.course.domain.repository.RestaurantCourseMenuRepository;
+import com.michelet.restaurantservice.course.domain.repository.RestaurantCourseRepository;
+import com.michelet.restaurantservice.domain.repository.RestaurantRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
